requirements.txt
accelerate==1.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
blis==1.0.1
botocore==1.37.34
Brotli @ file:///croot/brotli-split_1714483155106/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpathlib==0.20.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
diffusers==0.31.0
distro==1.9.0
einops==0.8.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.26.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.36.0
importlib_metadata==8.5.0
iopath @ file:///home/conda/feedstock_root/build_artifacts/iopath_1723846728822/work
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
multidict==6.1.0
murmurhash==1.0.10
networkx==3.4.2
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
omegaconf==2.3.0
openai==0.28.1
opencv-python==4.10.0.84
packaging==24.1
pillow @ file:///croot/pillow_1721059439630/work
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1728931699871/work
preshed==3.0.9
propcache==0.2.0
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytorch3d==0.7.5
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757090803/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rich==13.9.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1665138452165/work
tenacity==9.1.2
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1728288976871/work
thinc==8.3.2
tifffile==2024.9.20
timm==0.6.7
tokenizers==0.20.1
torch==1.13.0
torchvision==0.14.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1730145186354/work
transformers==4.46.1
typer==0.12.5
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
yarl==1.17.1
zipp==3.20.2
